Dates related functions

Dates related functions

Functions

FinancialToolbox.fromExcelNumberToDate — Function.

From Excel Number Format to Date

Date=fromExcelNumberToDate(ExcelNumber)

Where:

ExcelNumber = Integer representing a date in the excel format.

	Date      = Date representing the input in the Julia object format.

Example

julia> fromExcelNumberToDate(45000)
2023-03-15

source

FinancialToolbox.daysact — Function.

Actual Number of days between two dates

ndays=daysact(Date1,Date2)

Where:

Date1 = Start date.
	Date2 = End date.

	ndays      = Actual Number of days between Start Date and End Date.

Example

julia> daysact(Date(1996,10,12),Date(1998,1,10))
455

source

FinancialToolbox.yearfrac — Function.

Fraction of year between two Dates according the following convention

yfr=yearfrac(Date1,Date2,basis)

Where:

Date1 = Start date.
	Date2 = End date.
	basis = Integer representing the following conventions:
			- 0 = (ACT/ACT)
			- 1 = (30/360 SIA)
			- 2 = (ACT/360)
			- 3 = (ACT/365)
			- 4 = (30/360 PSA)
			- 5 = (30/360 ISDA)
			- 6 = (30E/360)
			- 7 = (ACT/365 JPN)

	yfr      = fraction of year between start and end date according to basis.

Example

julia> yearfrac(Date(1996,10,12),Date(1998,1,10),1)
1.2444444444444445

source

Index

FinancialToolbox.blkimpv

FinancialToolbox.blkprice

FinancialToolbox.blsdelta

FinancialToolbox.blsgamma

FinancialToolbox.blsimpv

FinancialToolbox.blslambda

FinancialToolbox.blsprice

FinancialToolbox.blspsi

FinancialToolbox.blsrho

FinancialToolbox.blstheta

FinancialToolbox.blsvanna

FinancialToolbox.blsvega

FinancialToolbox.daysact

FinancialToolbox.fromExcelNumberToDate

FinancialToolbox.normcdf

FinancialToolbox.normpdf

FinancialToolbox.yearfrac